totalCost type changed to BigDecimal

+++ src/main/java/pl/touk/ticketBooking/domain/bill/Bill.java
@@ -4,17 +4,19 @@ import lombok.Builder;
 import lombok.Getter;
 import pl.touk.ticketBooking.domain.guest.Guest;
 import pl.touk.ticketBooking.domain.ticket.Ticket;
+
+import java.math.BigDecimal;
 import java.time.LocalTime;
 
 @Getter
 @Builder
 public class Bill {
-    private float totalCost;
+    private BigDecimal totalCost;
     private LocalTime reservationExpirationTime;
 
     public static Bill createBill(Guest guest, LocalTime sessionTime){
         return Bill.builder()
-                .totalCost((float) guest.getTickets().stream().mapToDouble(Ticket::getPrice).sum())
+                .totalCost(BigDecimal.valueOf(guest.getTickets().stream().mapToDouble(Ticket::getPrice).sum()))
                 .reservationExpirationTime(sessionTime)
                 .build();
